+++ tests/testthat/test-slice.R
@@ -26,7 +26,7 @@ test_that("doesn't return NAs", {
   expect_equal(sliced_df, test_df)
 })
 
-test_that("works with irregular gaps & doesn't return NAs", {
+test_that("works with gaps & doesn't return NAs", {
   test_df <- tidytable(x = c(1,2,3,4), y = c(4,5,6,7), z = c("a", "a", "a", "b"))
 
   sliced_df <- test_df %>%
@@ -79,6 +79,7 @@ test_that("works with .by", {
   sliced_df <- test_df %>%
     slice.(1, .by = z)
 
+  expect_named(sliced_df, names(test_df))
   expect_equal(sliced_df$z, c("a", "b"))
   expect_equal(sliced_df$x, c(1, 4))
   expect_equal(sliced_df$y, c(4, 7))
@@ -166,13 +167,16 @@ test_that("_head.() works with n specified", {
   expect_equal(sliced_df, head(test_df, 3))
 })
 
-test_that("_head.() works with n specified", {
+test_that("_head.() works with n specified with by", {
   test_df <- tidytable(x = 1:10, y = 20:11, z = c(rep("a", 6), rep("b", 4)))
 
-  datatable_df <- test_df[, head(.SD, 3), by = z]
+  col_order <- names(test_df)
+
+  datatable_df <- test_df[, head(.SD, 3), by = z][, ..col_order]
   sliced_df <- test_df %>%
     slice_head.(n = 3, .by = z)
 
+  expect_named(sliced_df, names(test_df))
   expect_equal(datatable_df, sliced_df)
 })
 
@@ -217,7 +221,7 @@ test_that("_tail.() works with n specified", {
   expect_equal(sliced_df, tail(test_df, 3))
 })
 
-test_that("_tail.() works with n specified", {
+test_that("_tail.() works with n.()", {
   test_df <- tidytable(x = 1:10, y = 20:11, z = c(rep("a", 6), rep("b", 4)))
   sliced_df <- test_df %>%
     slice_tail.(n = n.())
@@ -228,10 +232,13 @@ test_that("_tail.() works with n specified", {
 test_that("_tail() works with n specified with .by", {
   test_df <- tidytable(x = 1:10, y = 20:11, z = c(rep("a", 6), rep("b", 4)))
 
-  datatable_df <- test_df[, tail(.SD, 3), by = z]
+  col_order <- names(test_df)
+
+  datatable_df <- test_df[, tail(.SD, 3), by = z][, ..col_order]
   sliced_df <- test_df %>%
     slice_tail.(n = 3, .by = z)
 
+  expect_named(sliced_df, names(test_df))
   expect_equal(datatable_df, sliced_df)
 })
 
@@ -274,6 +281,7 @@ test_that("_min.() works with .by", {
   sliced_df <- test_df %>%
     slice_min.(order_by = x, n = 3, .by = z)
 
+  expect_named(sliced_df, names(test_df))
   expect_equal(sliced_df$z, c("a", "a", "a", "b", "b", "b"))
   expect_equal(sliced_df$y, c(20,19,18,14,13,12))
 })