+++ tests/testthat/test-mutate_variants.R
@@ -138,6 +138,38 @@ test_that("mutate_across() works with newly named columns", {
   expect_equal(df$y_new, c(3,3,3))
 })
 
+test_that("mutate_across() works with newly named columns using .names with single .fn", {
+  df <- data.table(x = c(1,1,1), y = c(2,2,2), z = c("a", "a", "b"))
+  df <- df %>%
+    mutate_across.(c(x:y), ~ .x + 1, .names = "new_{col}")
+
+  expect_named(df, c("x","y","z","new_x","new_y"))
+  expect_equal(df$new_x, c(2,2,2))
+  expect_equal(df$new_y, c(3,3,3))
+})
+
+test_that("mutate_across() works with newly named columns using .names", {
+  df <- data.table(x = c(1,1,1), y = c(2,2,2), z = c("a", "a", "b"))
+  df <- df %>%
+    mutate_across.(c(x:y), list(new = function(.x) .x + 1), .names = "{fn}_{col}")
+
+  expect_named(df, c("x","y","z","new_x","new_y"))
+  expect_equal(df$new_x, c(2,2,2))
+  expect_equal(df$new_y, c(3,3,3))
+})
+
+test_that("mutate_across() works with newly named columns using .names w/ autonaming", {
+  df <- data.table(x = c(1,1,1), y = c(2,2,2), z = c("a", "a", "b"))
+  df <- df %>%
+    mutate_across.(c(x:y), list(new = ~ .x + 1, ~ .x + 2), .names = "{col}_{fn}_stuff")
+
+  expect_named(df, c("x","y","z","x_new_stuff","y_new_stuff", "x_1_stuff", "y_1_stuff"))
+  expect_equal(df$x_new_stuff, c(2,2,2))
+  expect_equal(df$y_new_stuff, c(3,3,3))
+  expect_equal(df$x_1_stuff, c(3,3,3))
+  expect_equal(df$y_1_stuff, c(4,4,4))
+})
+
 # twiddle testing ----------------------------
 test_that("mutate_if.() works with twiddle", {
   df <- data.table(x = c(1,1,1), y = c(2,2,2), z = c("a", "a", "b"))