+++ NEWS.md
@@ -1,3 +1,5 @@
+# tidytable 0.5.4.9 (in development)
+
 # tidytable 0.5.4
 
 #### Breaking changes